move vitest config files to packages/core
