+++ packages/core/package.json
@@ -72,7 +72,7 @@
     "start": "stencil build --dev --watch --serve",
     "test.setup": "playwright install chromium --with-deps",
     "test": "npm run test.spec && npm run test.ct",
-    "test.spec": "vitest run --config src/vitest.config.ts --project spec",
+    "test.spec": "vitest run --config vitest.config.ts --project spec",
     "test.ct": "playwright test --config playwright.config.ts --reporter list",
     "generate": "stencil generate",
     "host-root": "http-server -a 127.0.0.1 -p 8080 -c-1"
