+++ packages/core/tsconfig.json
@@ -3,6 +3,7 @@
   "references": [
     { "path": "./tsconfig.lib.json" },
     { "path": "./tsconfig.node.json" },
+    { "path": "./tsconfig.vitest.json" },
     { "path": "./tsconfig.test.json" }
   ]
 }
